Fixed up for packaging and added Swift demo app. (#12)
* Fixed up for packaging and added Swift demo app.
This starts to add packaging logic into the javascript
and kotlin build scripts. As a part of that, the bundled
licensing files needed to get duplicated in each subdirectory so that
they can be used by each bundler.
Also, because the README.md files are split, they've now been
modified to be appropriate for each language.
Finally, this adds a demo app for Swift and fixes the Swift
code so that it correctly exports the API and produces valid
GeoJSON.

* Addressed PR feedback.
Fixed up CONTRIBUTING.md links, changed unpkg to jsdelivr.